requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
pydantic==2.6.4
pyjwt==2.8.0
cryptography==42.0.5
pyodbc==5.1.0
python-dotenv==1.0.1
sqlalchemy==2.0.29
bcrypt==4.1.2
python-dotenv==1.0.1
pypdf==4.2.0
azure-storage-blob==12.19.1
asyncio==3.4.3
asyncpg==0.29.0
greenlet==3.0.3
